backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.0.0
httpx==0.28.1